requirements.txt
sqlalchemy[asyncio]
psycopg2-binary
pandas
polars
asyncpg
aiofiles
aiocsv
pyarrow